Move files and rename components
